+++ src/main/java/com/peter/financeapp/model/User.java
@@ -3,19 +3,19 @@ package com.peter.financeapp.model;
 import java.time.LocalDate;
 
 public class User {
-    private long id;
+    private Long id;
     private String username;
     private String password;
     private LocalDate createdAt;
 
-    public User (int id,String username,String password,LocalDate createdAt){
+    public User (Long id,String username,String password,LocalDate createdAt){
         this.id = id; this.username = username; this.password = password; this.createdAt = createdAt;
     }
     public User(String username,String password){
         this.username = username; this.password = password;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -31,7 +31,7 @@ public class User {
         this.createdAt = createdAt;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
